validate witness cmd
