+++ chain/chain/src/lib.rs
@@ -19,7 +19,7 @@ mod doomslug;
 pub mod flat_storage_creator;
 mod garbage_collection;
 mod lightclient;
-mod metrics;
+pub mod metrics;
 pub mod migrations;
 pub mod missing_chunks;
 pub mod orphan;
@@ -27,7 +27,7 @@ pub mod resharding;
 pub mod runtime;
 mod state_request_tracker;
 pub mod state_snapshot_actor;
-mod stateless_validation;
+pub mod stateless_validation;
 mod store;
 pub mod store_validator;
 pub mod test_utils;
