+++ chain/chain/src/stateless_validation/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod chunk_endorsement;
 pub(crate) mod state_transition_data;
+pub mod state_witness;
